feat: 역 삭제 시 없는 역을 삭제하려고 하면 예외를 발생하도록 하는 기능 추가

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -3,6 +3,7 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.StationDao;
@@ -39,6 +40,9 @@ public class StationService {
     }
 
     public void delete(Long id) {
+        if (showAll().isEmpty()) {
+            throw new EmptyResultDataAccessException("존재하지 않는 역입니다.", 1);
+        }
         stationDao.deleteById(id);
     }
